chore: resolve clippy warnings for -D warnings CI gate
Clears ~300 clippy warnings across 89 files so the CI clippy job
(cargo clippy --all --all-targets -- -D warnings) goes green.

Mechanical cleanup categories (mostly auto-applied via cargo clippy --fix):
unnecessary_map_or, manual_is_multiple_of, manual_strip, new_without_default,
field_reassign_with_default, doc_lazy_continuation, vec_init_then_push,
redundant_closure, collapsible_if, etc.

Non-mechanical fixes:
- claude-tui/src/app.rs: restructure the post-slash-command side-effect
  block so the std::sync::Mutex guard on shared_state is dropped before
  the engine_tx.send(...).await call (clippy::await_holding_lock real
  hit, not a false positive). Same file: convert a never-looping while
  into an if.
- mcp/result_transform.rs: work around clippy::approx_constant on a
  literal 3.14 used as a type probe (not an actual PI).
- Test modules using std::sync::Mutex as a process-wide serializer for
  env/global-state tests get a module-scoped #[allow(await_holding_lock)]
  with a comment explaining why the lock is safe there.

+++ crates/claude-core/src/mcp/datetime_parser.rs
@@ -126,7 +126,7 @@ fn weekday_name(w: chrono::Weekday) -> &'static str {
 fn format_timezone<Tz: chrono::TimeZone>(dt: &chrono::DateTime<Tz>) -> String {
     let offset_secs = dt.offset().fix().local_minus_utc();
     let sign = if offset_secs >= 0 { '+' } else { '-' };
-    let abs = offset_secs.unsigned_abs() as u32;
+    let abs = offset_secs.unsigned_abs();
     let hours = abs / 3600;
     let mins = (abs % 3600) / 60;
     format!("{}{:02}:{:02}", sign, hours, mins)
@@ -181,7 +181,10 @@ mod tests {
 
     #[test]
     fn normalize_rejects_empty_and_invalid() {
-        assert!(matches!(normalize_model_reply(""), DateTimeParseResult::Err(_)));
+        assert!(matches!(
+            normalize_model_reply(""),
+            DateTimeParseResult::Err(_)
+        ));
         assert!(matches!(
             normalize_model_reply("INVALID"),
             DateTimeParseResult::Err(_)
